Refactor and optimize data processing module
Refactored data processing functions for better readability and performance. Optimized algorithm to reduce runtime complexity. Updated unit tests to cover new edge cases. Fixed minor bugs in data validation logic.
