+++ Core.Application/Features/Scraping/PelisPlusLat/Commands/GetPelisPlusLatMovies/GetPelisPlusLatMoviesCommand.cs
@@ -2,7 +2,7 @@
 using Core.Application.DTOs.Relations;
 using Core.Application.Helpers.TMDB;
 using Core.Application.Interface.Repositories;
-using Core.Domain.Entities.GeneralMovie;
+using Core.Domain.Entities.Movie;
 using Core.Domain.Entities.Relations;
 using Core.Domain.Entities.WebScraping;
 using MediatR;
