+++ Core.Domain/Entities/Relations/MovieList_Movie.cs
@@ -1,5 +1,4 @@
 ﻿using Core.Domain.Common;
-using Core.Domain.Entities.GeneralMovie;
 using Core.Domain.Entities.UserThings;
 
 namespace Core.Domain.Entities.Relations
