+++ Core.Domain/Entities/Relations/Movie_MovieWeb.cs
@@ -1,5 +1,4 @@
 ﻿using Core.Domain.Common;
-using Core.Domain.Entities.GeneralMovie;
 using Core.Domain.Entities.WebScraping;
 
 namespace Core.Domain.Entities.Relations
