+++ Core.Domain/Entities/UserThings/Recents.cs
@@ -1,6 +1,5 @@
 ﻿using Auth.Infraestructure.Identity.Entities;
 using Core.Domain.Common;
-using Core.Domain.Entities.GeneralMovie;
 
 namespace Core.Domain.Entities.UserThings
 {
