+++ Infraestructure.Persistence/Repositories/MovieWebRepository.cs
@@ -1,10 +1,10 @@
 ﻿using Core.Application.DTOs.Scraping;
 using Core.Application.Interface.Repositories;
 using Core.Domain.Entities.WebScraping;
-using Infraestructure.Persistence.Context;
+using Infrastructure.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 
-namespace Infraestructure.Persistence.Repositories
+namespace Infrastructure.Persistence.Repositories
 {
     public class MovieWebRepository : GenericRepository<MovieWeb>, IMovieWebRepository
     {
