+++ Infraestructure.Persistence/Repositories/Movie_MovieWebRepository.cs
@@ -1,8 +1,8 @@
 ﻿using Core.Application.Interface.Repositories;
 using Core.Domain.Entities.Relations;
-using Infraestructure.Persistence.Context;
+using Infrastructure.Persistence.Context;
 
-namespace Infraestructure.Persistence.Repositories
+namespace Infrastructure.Persistence.Repositories
 {
     public class Movie_MovieWebRepository : GenericRepository<Movie_MovieWeb>, IMovie_MovieWebRepository
     {
