+++ Infraestructure.Persistence/Repositories/RecentsRepository.cs
@@ -1,8 +1,8 @@
 ﻿using Core.Application.Interface.Repositories;
 using Core.Domain.Entities.UserThings;
-using Infraestructure.Persistence.Context;
+using Infrastructure.Persistence.Context;
 
-namespace Infraestructure.Persistence.Repositories
+namespace Infrastructure.Persistence.Repositories
 {
     public class RecentsRepository : GenericRepository<Recents>, IRecentsRepository
     {
